Last commit

+++ Assignment 3/Paltineanu Rares - Mihai/frontend V2.0/frontend/src/app/user/admin/components/users/toolbar/toolbar.component.ts
@@ -32,6 +32,11 @@ export class UserToolbarComponent {
   }
 
   public onDeleteUser(): void {
+    let user = JSON.parse(localStorage.getItem('user')!);
+    if(this.grid.selectedRowKeys[0].username === user.name) {
+      alert('You can not delete yourself')
+      return
+    }
     this.userService.deleteUser(this.grid.selectedRowKeys[0].id).subscribe(() => {
       this.updateUsersGrid()});
     ;
